+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/config/controller/BoardsController.kt
@@ -1,6 +1,7 @@
-package com.cognifide.cogboard.config
+package com.cognifide.cogboard.config.controller
 
 import com.cognifide.cogboard.CogboardConstants
+import com.cognifide.cogboard.config.EndpointLoader
 import com.cognifide.cogboard.storage.Storage
 import com.cognifide.cogboard.storage.docker.VolumeStorage
 import com.cognifide.cogboard.widget.Widget
@@ -10,16 +11,14 @@ import io.vertx.core.json.JsonObject
 import io.vertx.core.logging.Logger
 import io.vertx.core.logging.LoggerFactory
 
-class WidgetsController : AbstractVerticle() {
+internal open class BoardsController : AbstractVerticle() {
 
-    private val widgets = mutableMapOf<String, Widget>()
-    private lateinit var storage: Storage
+    internal lateinit var storage: Storage
+    internal val widgets = mutableMapOf<String, Widget>()
 
     override fun start() {
         storage = VolumeStorage(vertx)
         listenOnBoardsConfigSave()
-        listenOnWidgetUpdate()
-        listenOnWidgetDelete()
         loadBoardsConfig()
     }
 
@@ -27,17 +26,6 @@ class WidgetsController : AbstractVerticle() {
             .eventBus()
             .consumer<JsonObject>(CogboardConstants.EVENT_SAVE_BOARDS_CONFIG)
             .handler { storage.saveBoardsConfig(it.body()) }
-
-    private fun listenOnWidgetUpdate() = vertx
-            .eventBus()
-            .consumer<JsonObject>(CogboardConstants.EVENT_UPDATE_WIDGET_CONFIG)
-            .handler { createOrUpdate(it.body()) }
-
-    private fun listenOnWidgetDelete() = vertx
-            .eventBus()
-            .consumer<JsonObject>(CogboardConstants.EVENT_DELETE_WIDGET_CONFIG)
-            .handler { delete(it.body()) }
-
 
     private fun loadBoardsConfig() = storage
             .loadBoardsConfig()
@@ -47,7 +35,7 @@ class WidgetsController : AbstractVerticle() {
                 createOrUpdate(JsonObject(it.value.toString()))
             }
 
-    private fun createOrUpdate(config: JsonObject) {
+    fun createOrUpdate(config: JsonObject) {
         var newConfig = config
         val id = config.getString(CogboardConstants.PROP_ID)
 
@@ -63,16 +51,6 @@ class WidgetsController : AbstractVerticle() {
         }
     }
 
-    private fun delete(config: JsonObject) {
-        val id = config.getString(CogboardConstants.PROP_ID)
-
-        if (id != null) {
-            widgets.remove(id)?.stop()
-        } else {
-            LOGGER.error("Widget Delete | There is widget with no ID in configuration: $config")
-        }
-    }
-
     private fun attachEndpoint(config: JsonObject) {
         val endpointId = config.getString(CogboardConstants.PROP_ENDPOINT)
         endpointId?.let {
@@ -82,6 +60,6 @@ class WidgetsController : AbstractVerticle() {
     }
 
     companion object {
-        val LOGGER: Logger = LoggerFactory.getLogger(WidgetsController::class.java)
+        val LOGGER: Logger = LoggerFactory.getLogger(BoardsController::class.java)
     }
 }
