+++ knotx/conf/cogboard.conf
@@ -1,7 +1,7 @@
 modules {
   websocketsServer = "com.cognifide.cogboard.http.WebSocketsServer"
-  endpointsController = "com.cognifide.cogboard.config.EndpointsController"
-  widgetsController = "com.cognifide.cogboard.config.WidgetsController"
+  endpointsController = "com.cognifide.cogboard.config.controller.EndpointsController"
+  boardsController = "com.cognifide.cogboard.config.controller.BoardsController"
   httpClient = "com.cognifide.cogboard.http.HttpClient"
 }
 
@@ -11,11 +11,11 @@ config {
     port = $ws_port
   }
 
-  endpointsManager.options.config {
+  endpointsController.options.config {
     include required(file("/data/endpoints.json"))
   }
 
-  widgetsManager.options.config {
+  boardsController.options.config {
     include required(file("/data/endpoints.json"))
   }
